Add docker.io Drone plugin which creates a docker image for each successful build
This change adds a new Drone plugin for creating a docker image. The plugin
creates a docker image for the test build and pushes the image to images.docker.io.

I've just tested it end-to-end with 'drone build' from the command line because
I'm not sure of another test that would be useful.