Allow external environment variables
Pass any environment variables prefixed with ``DRONE_`` into the build
script that's executed in the container. For example, the following
command would export an environment variable called ``BUILDER`` with a
value of ``node-01`` in the build script:

    DRONE_BUILDER=node-01 drone build

The intent is to provide allow a little more insight into the build host
environment should a build break.

This is the same code from #508 but for the ``exp`` branch.