Pin NPM pkgs with shrinkwrap for repeatable builds
Pin NPM packages with `npm shrinkwrap --dev`.

Adds a minimal `package.json` created by running `npm init` and removing
some lines; just enough to satistify `npm install` without it showing
warnings.

Add the `node_modules/` directory to the `.gitignore` file as we can
probably rely on the NPM registry server and this diverges as little as
possible from the current behaviour (i.e. don't vendor anything). The
decision not to vendor the NPM modules can always be reviewed later.

This commit tells NPM to install the packages in the current directory
rather than globally, which will make it easier to vendor the NPM
packages in future if desired.