Change org name in GitHub testdata
The example JSON taken from GitHub's API documentation doesn't indicate that
`login` is the name of an organisation rather than a user. Change it to
something that looks more like an org, because it will make a test that I'm
about to add more readable. The endpoint name changes accordingly.