Make build runner configurable/pluggable
Closes #1324

This change adds support for configuring the build, pull-request, and
notify agent containers via the BUILD_IMAGE, PULL_REQUEST_IMAGE, and
NOTIFY_IMAGE environment variables. The following inline YAML format is
supported:

    BUILD_IMAGE="image: drone/drone-exec, entrypoint: [/bin/drone-exec], cmd: [--clone, --cache, --build, --publish, --deploy], env: [FOO=bar, THIS=that]"

All of the keys are optional. If a key is omitted then the default
value will be used.

If there is an error parsing YAML then the application will panic on
startup.