BUILDING: add export CGO_ENABLED=0
While this will work fine without it, you'll hit an issue when you go to
run a docker container built with this because the resulting binary will
look for some dynamic library on start and fail to find it which yields
a "no such file or directory" error.

This at least hints someone down the right track.  If you're using
sqlite3 locally, then you'll need CGO enabled (AFAICT).