add working docker-compose
Using drone is easy, when providing a simple docker-compose setup.
However, the documentation is outdated and the information on the
net like stackoverflow are wrong in terms of producing errors in the
most recent versions, where the agent cannot talk to the server
using GRPC.

This removes the port 9000 to be not exposed to the public (being
just open for internal connections) and the link from the agent ot
server is added.

I hope this solves all the confusion of how to use drone, as it is
now a simple "docker-compose up".