agent: speed up trimming of logged lines
Scanning through a buffer should be much faster
than encoding JSON every time through the loop.

Looking for '"},' isn't a perfect approximation of the next line,
but it should be good enough to keep us synchronized.

This should reduce the calls to json.Marsal on a large array
of lines quite considerably.