Add req.Build.Event to the cache's key. #CI-98
If the key contains only repo ID and SHA commit, then it leads to missing events, because for different event types (`push`, `pull_request`) that key is the same. For example, when the first event comes in (for instance, `push` event), it saves in the cache. When another event, for instance, `pull_request` comes in, it doesn't processing since it already in the cache (because the key is the same).