Improve the logging for event matching
While developing:

  https://github.com/drone/drone/pull/2902

It was difficult to verify which events the server was receiving and
how. This commit was applied to the server to make transparent what
events the solver was solving for and immediately cleared up the issue.

It stands to reason it might be useful for other users debugging their
pipelines.