Implement PR "Close" events
There is currently a feature request associated with being able to run
some sort of pipeline at the termination of some git event; specifically
in this case a "closure" of a pull request.

This allows the consumers of drone to use drone to spin up ephemeral
environments for the purpose of testing, and then later have them torn
down automatically by Drone once the pull request has been closed.

This commit implements such a solution after reading the guidance on:

  https://github.com/drone/drone/pull/2884

It attempts to use the "actions" component of the spec rather than
create new first level event primitive, but stubs out that actions
primitive in the case that the user has not supplied any so as to
preserve BC with the existing pipelines.

== Design Notes
=== General Design

The design follows the patn of other drone events as much as possible.
Because this is the first event that would likely break users (unlike
the "sync" event on pull requests for example), BC preservation is a
little more difficult. If this comes up frequently in future it might be
reasonable to extend a "default "pipeline object, rather than
deserialize the pipeline only from Yaml.

== Test Plan

This was tested with the following .drone.yml:

  ---
  kind: "pipeline"
  type: "kubernetes"
  name: "push"
  steps:
    - name: "container"
      image: "busybox"
      commands:
        - echo "Pushed"
  trigger:
    event:
      include:
        - pull_request
  ---
  kind: "pipeline"
  type: "kubernetes"
  name: "closed"
  clone:
    disable: true
  steps:
    - name: "container"
      image: "busybox"
      commands:
        - echo "Closed"
  trigger:
    event:
      include:
        - pull_request
    action:
      include:
        - closed

Which worked with:

1. Creating the PR
2. Updating the PR
3. Closing the PR
4. Merging the PR

Additionally tested with the events deliberately specified for the
pushed event:

- created
- synchronized

in the "push" type.