Rework the solution for PR close events
In the previous design the pipeline as supplied by the user was modified
before passing it to the skip events, injecting a set of defaults where
none existed. However, that does not follow the patterns otherwise found
in the codebase and with the primary author and thus makes the code
more difficult to reason of.

This commit reimplements the functionality as a set of "skip"
functionality. It introduces:

- `skipEventAction`:      The primitive for conditional skipping based on
                          event/action combinations
- `skipPullRequestEval`:  The evaluation set for the pull request

This is hopefully clearer to reason through, but still allows the pull
request close events.

== Design Notes
=== No unit tests

This work was done and tested quickly, and the author just tested the
pipeline by running builds as part of their work.

It should have unit tests, but the author does not have the time
currently to contribute them.