Fix Gitea branch builds
Drone prefixes a branch build ref with `ref/heads/`, which results in
go-scm constructing a URL to Gitea's commits API as
`.../commits/ref%2Fheads%2fmy-branch`, which gets a 404 response from
Gitea. Removing the prefix and only using `my-branch` for the full ref
now correctly has drone getting the commit details for the branch and
builds are built successfully.

Tested with Gitea 1.13.6 (released March 24, 2021).